Moved final file and removed old view files
