--- src/views/gameGui/gameGui.css
@@ -1,16 +1,0 @@
-#gameGui {
-  top: 5vw;
-  text-align: center;
-}
-
-.endGameMessage {
-  background-color: rgba(0, 128, 128, 0.2);
-  border: 1px solid white;
-  padding: 1vw;
-  display: inline-block;
-  border-radius: 3px;
-}
-
-.endGameText {
-  margin-bottom: 1vw;
-}
