--- src/views/mainMenu/mainMenu.css
@@ -1,51 +1,0 @@
-.mainMenu {
-  text-align: center;
-  top: 17vw;
-  bottom: 0;
-}
-
-.titleText {
-  font-size: 10vw;
-}
-
-.mainMenuButton {
-  font-size: 2vw;
-  display: flex;
-  align-items: center;
-}
-
-.mainMenuButton:hover {
-  cursor: pointer;
-  color: gray;
-}
-
-.mainMenuButton .icon{
-  display: inline;
-  opacity: 1;
-}
-
-.mainMenuButton-puzzle{
-  position: absolute;
-  bottom: 5vw;
-  left: 30vw;
-}
-
-.mainMenuButton-puzzle .icon{
-  margin-right: 1vw;
-}
-
-.mainMenuButton-free{
-  position: absolute;
-  bottom: 5vw;
-  left: 54vw;
-}
-
-.mainMenuButton-free svg {
-  margin-right: 2.5vw;
-  transform: rotate(30deg);
-}
-
-.mainMenuButton svg{
-  width: 7vw;
-  height: 7vw;
-}
