--- src/views/mapEditor/mapEditor.css
@@ -1,27 +1,0 @@
-#mapEditor {
-  position: absolute;
-  margin: 1vw;
-  border-radius: 3px;
-  color: white;
-  text-align: center;
-  font-size: 1.2vw;
-  background-color: rgba(0, 128, 128, 0.2);
-  border: 1px solid white;
-  top: 30vw;
-}
-
-#mapEditor > * {
-  padding: 1vw;
-}
-
-#mapEditor > * + * {
-  border-top: 1px solid white;
-}
-
-.mapEditorSteps > * {
-  display: block;
-}
-
-.mapEditorSteps > * + * {
-  margin-top: 1vw;
-}
