+++ src/vues/player.js
